core: organize reduceC SIMD backends
